Patient Status Display page now functional
